move to legacy: pyfunc_type_util.py
