+++ weave/legacy/decorator_op.py
@@ -4,9 +4,9 @@ from typing import Callable, Optional, TypeVar
 
 from typing_extensions import ParamSpec
 
-from weave import pyfunc_type_util, registry_mem
+from weave import registry_mem
 from weave import weave_types as types
-from weave.legacy import context_state, derive_op, op_args, op_def
+from weave.legacy import context_state, derive_op, op_args, op_def, pyfunc_type_util
 
 if typing.TYPE_CHECKING:
     from weave.legacy.gql_op_plugin import GqlOpPlugin
