+++ weave/legacy/pyfunc_type_util.py
@@ -3,8 +3,8 @@ import typing
 
 from weave.legacy import infer_types, op_args
 
-from . import errors
-from . import weave_types as types
+from .. import errors
+from .. import weave_types as types
 
 InputTypeItemType = typing.Union[types.Type, typing.Callable[..., types.Type]]
 InputTypeType = typing.Union[op_args.OpArgs, typing.Mapping[str, InputTypeItemType]]
